% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rray2-package.R
\docType{package}
\name{rray2-package}
\alias{rray2}
\alias{rray2-package}
\title{rray2: Simple Arrays}
\description{
Provides a toolkit for manipulating arrays in a consistent, 
    powerful, and intuitive manner through the use of broadcasting and a new
    array class, the 'rray'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/rray2}
  \item Report bugs at \url{https://github.com/DavisVaughan/rray2/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
