% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{rray_add}
\alias{rray_add}
\title{Arithmetic}
\usage{
rray_add(x, y)
}
\arguments{
\item{x, y}{Arrays}
}
\value{
An array with dimensions of \code{rray_dims_common(x, y)}.
}
\description{
These functions perform arithmetic between multiple arrays, with
broadcasting.
}
\examples{
rray_add(matrix(1:5, nrow = 1), matrix(1:5, ncol = 1))
}
