% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality.R
\name{rray_dimensionality}
\alias{rray_dimensionality}
\title{Compute dimensionality}
\usage{
rray_dimensionality(x)
}
\arguments{
\item{x}{A vector}
}
\description{
\code{rray_dimensionality()} computes the dimensionality of an object. This is
also known as the "number of dimensions".
}
\section{Invariants}{

\itemize{
\item \code{vec_size(rray_dimensionality(x)) == 1L}
\item \code{vec_ptype(rray_dimensionality(x)) == integer()}
\item \code{rray_dimensionality(x) == vec_size(rray_dims(x))}
}
}

\examples{
rray_dimensionality(1)
rray_dimensionality(matrix())
rray_dimensionality(NULL)
}
