% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{rray_dims}
\alias{rray_dims}
\title{Compute dimensions}
\usage{
rray_dims(x)
}
\arguments{
\item{x}{A vector}
}
\description{
\code{rray_dims()} computes the dimensions of a vector.

This is similar to \code{\link[=dim]{dim()}}, but errors on scalar objects and never
returns \code{NULL}. The dimensions of a bare vector like \code{1:5} are its length.
}
\examples{
rray_dims(1)
rray_dims(array(1:6, dim = c(2, 3)))

try(rray_dims(mean))
}
