% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{rray_elements}
\alias{rray_elements}
\title{Number of elements}
\usage{
rray_elements(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A single double. The number of elements in \code{x}.
}
\description{
\code{rray_elements()} computes the number of individual elements in a vector.
}
\details{
It generally computes the same thing as \code{length()}, but has a more
predictable name and errors on scalar objects.
}
\section{Invariants}{

\itemize{
\item \code{vec_size(rray_elements(x)) == 1L}
\item \code{rray_elements(x) == prod(rray_dims(x))}
}
}

\examples{
rray_elements(1:5)

rray_elements(matrix(1, 2, 2))

# It is different from `vec_size()`,
# which returns the number of observations
library(vctrs)
vec_size(matrix(1, 2, 2))
}
